tests/wpt/tests/css/css-lists/parsing/counter-set-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Lists: getComputedStyle().counterSet</title>
<link rel="help" href="https://drafts.csswg.org/css-lists-3/#propdef-counter-set">
<meta name="assert" content="tests that counter-set grammar is supported.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('counter-set', 'none');
test_computed_value('counter-set', 'myCounter', 'myCounter 0');
test_computed_value('counter-set', 'myCounter 5');
test_computed_value('counter-set', 'myCounter 7 otherCounter 8');
test_computed_value('counter-set', 'myCounter otherCounter 8', 'myCounter 0 otherCounter 8');
test_computed_value('counter-set', 'myCounter 7 otherCounter', 'myCounter 7 otherCounter 0');
</script>
</body>
</html>
